Rename pit class and add logs
