+++ legacy/src/main/java/org/opensearch/sql/legacy/executor/join/ElasticJoinExecutor.java
@@ -39,8 +39,8 @@ import org.opensearch.sql.legacy.domain.Field;
 import org.opensearch.sql.legacy.esdomain.LocalClusterState;
 import org.opensearch.sql.legacy.exception.SqlParseException;
 import org.opensearch.sql.legacy.executor.ElasticHitsExecutor;
-import org.opensearch.sql.legacy.pit.PIT;
 import org.opensearch.sql.legacy.pit.PointInTimeHandler;
+import org.opensearch.sql.legacy.pit.PointInTimeHandlerImpl;
 import org.opensearch.sql.legacy.query.SqlElasticRequestBuilder;
 import org.opensearch.sql.legacy.query.join.HashJoinElasticRequestBuilder;
 import org.opensearch.sql.legacy.query.join.JoinRequestBuilder;
@@ -59,7 +59,7 @@ public abstract class ElasticJoinExecutor implements ElasticHitsExecutor {
   private boolean allFieldsReturn;
   protected Client client;
   protected String[] indices;
-  protected PIT pit;
+  protected PointInTimeHandler pit;
 
   protected ElasticJoinExecutor(Client client, JoinRequestBuilder requestBuilder) {
     metaResults = new MetaSearchResult();
@@ -101,7 +101,7 @@ public abstract class ElasticJoinExecutor implements ElasticHitsExecutor {
     try {
       long timeBefore = System.currentTimeMillis();
       if (LocalClusterState.state().getSettingValue(SQL_PAGINATION_API_SEARCH_AFTER)) {
-        pit = new PointInTimeHandler(client, indices);
+        pit = new PointInTimeHandlerImpl(client, indices);
         pit.create();
       }
       results = innerRun();
